+++ PFF/src/engine/render/frame_info.h
@@ -9,7 +9,7 @@ namespace PFF {
 	struct frame_info {
 
 		u32 frame_index;
-		f32 frame_time;
+		f64 frame_time;
 		VkCommandBuffer command_buffer;
 		std::shared_ptr<camera> camera;
 		VkDescriptorSet global_descriptor_set;
